client/pyproject.toml
[tool.poetry]
name = "client"
version = "0.1.0"
description = "simple USB KVM client"
authors = ["wevsty <ty@wevs.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.13,>=3.9"
pyyaml = "^6.0.2"
loguru = "^0.7.2"
pywin32 = "^306"
pyside6-addons = "^6.7.3"
pyside6-essentials = "^6.7.3"
pyside6 = "^6.7.3"
pywinhook = "^1.6.2"
pyserial = "^3.5"
ch9329 = "^1.2.7"
nuitka = "^2.4.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
